Contact analysis: move the example to cookbook - remove accidentally included CoordinateNeighborSearch, - single test class in test_contacts.py, remove test_contacts_best.py,
